moves report notification logic into ReportManager.cs - shows notification only if user enabled the related settings entry
